Add routes for taproomNews and taproomEvents. Remove taproot component,route, and module files.
The 'Taproot' menu button should not link to a page called 'taproot'.
The menu button should be a drop-down with options to go to the 'news'
page or the 'taproot-events' page.

Removed all files associated with a 'taproot' link and added components,
routes, and modules to support 'news' and 'taproot-events' pages.